requirements.txt
kafka-python
pymongo
requests
